refactoring to add tabs to role page #29949
