+++ projects/safe/src/i18n/test.json
@@ -700,6 +700,7 @@
 				"roles": {
 					"application": "******",
 					"backoffice": "******",
+					"groups": "******",
 					"selectedApplication": "******",
 					"selectedRoles": "******"
 				},
