+++ projects/safe/src/lib/components/roles/back-office-roles/back-office-roles.module.ts
@@ -1,0 +1,53 @@
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { MatButtonModule } from '@angular/material/button';
+import { MatDialogModule } from '@angular/material/dialog';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatIconModule } from '@angular/material/icon';
+import { MatInputModule } from '@angular/material/input';
+import { MatMenuModule } from '@angular/material/menu';
+import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
+import { MatSelectModule } from '@angular/material/select';
+import { MatTableModule } from '@angular/material/table';
+import { MatSortModule } from '@angular/material/sort';
+import { MatAutocompleteModule } from '@angular/material/autocomplete';
+import { MatDividerModule } from '@angular/material/divider';
+import { TranslateModule } from '@ngx-translate/core';
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { SafeConfirmModalModule } from '../../confirm-modal/confirm-modal.module';
+import { SafeButtonModule } from '../../ui/button/button.module';
+import { SafeSkeletonTableModule } from '../../skeleton/skeleton-table/skeleton-table.module';
+import { SafeBackOfficeRolesComponent } from './back-office-roles.component';
+import { SafeAddRoleComponent } from './components/add-role/add-role.component';
+
+/**
+ * BackOfficeRolesModule manages modules and components
+ * related to the back-office roles tab
+ */
+@NgModule({
+  declarations: [SafeBackOfficeRolesComponent, SafeAddRoleComponent],
+  imports: [
+    CommonModule,
+    CommonModule,
+    FormsModule,
+    ReactiveFormsModule,
+    MatFormFieldModule,
+    MatInputModule,
+    MatSelectModule,
+    MatProgressSpinnerModule,
+    MatTableModule,
+    MatDialogModule,
+    MatButtonModule,
+    MatMenuModule,
+    MatIconModule,
+    SafeConfirmModalModule,
+    MatSortModule,
+    MatAutocompleteModule,
+    MatDividerModule,
+    SafeButtonModule,
+    TranslateModule,
+    SafeSkeletonTableModule,
+  ],
+  exports: [SafeBackOfficeRolesComponent],
+})
+export class SafeBackOfficeRolesModule {}
