+++ projects/safe/src/lib/components/roles/groups/groups.component.html
@@ -1,0 +1,1 @@
+<p>groups works!</p>
