+++ projects/safe/src/lib/components/roles/groups/groups.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+/**
+ * This component is used to display the groups tab in the platform
+ */
+@Component({
+  selector: 'safe-groups',
+  templateUrl: './groups.component.html',
+  styleUrls: ['./groups.component.scss'],
+})
+export class SafeGroupsComponent implements OnInit {
+  constructor() {}
+
+  ngOnInit(): void {}
+}
