+++ projects/safe/src/lib/components/roles/groups/groups.module.ts
@@ -1,0 +1,14 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { SafeGroupsComponent } from './groups.component';
+
+/**
+ * SafeGroupsModule manages modules and components
+ * related to the groups tab
+ */
+@NgModule({
+  declarations: [SafeGroupsComponent],
+  imports: [CommonModule],
+  exports: [SafeGroupsComponent],
+})
+export class SafeGroupsModule {}
